requirements.txt
python-dotenv
anthropic
fastapi
uvicorn
supabase
dacite
pydantic